Automatically backed up by Learn

+++ calculator.rb
@@ -1,3 +1,3 @@
 # Add your variables here
-first_number = 2.to_f
+first_number = 2.to_i
 p first_number
